Update Readme page with relevant info, links
Included a link to the issue page that I created this repo for, and also included some other relevant info.

The included images were created by me and I'm planning to release them as part of an icon set later this year, and so to avoid any future trouble whatsoever, I'm releasing these 2 images under the CC-BY-SA license here.